+++ backend/swpp/model/user.py
@@ -1,0 +1,5 @@
+from django.db import models
+from django.contrib.auth import get_user_model
+from annoying.fields import AutoOneToOneField
+
+User = get_user_model()
