+++ backend/swpp/serializer/user.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from swpp.models import User
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = ('id', 'username', 'profile')
