+++ backend/swpp/serializers.py
@@ -1,17 +1,3 @@
-from rest_framework import serializers
-from swpp.models import User, Profile, Tutor
-
-class UserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = ('id', 'username', 'profile')
-
-class ProfileSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Profile
-        fields = ('user', 'joined', 'tutor')
-
-class TutorSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Tutor
-        fields = ('profile', 'bio')
+from swpp.serializer.user import UserSerializer 
+from swpp.serializer.profile import ProfileSerializer 
+from swpp.serializer.tutor import TutorSerializer
