+++ backend/swpp/view/tutor.py
@@ -1,0 +1,12 @@
+from django.shortcuts import render
+from swpp.models import Tutor
+from swpp.serializers import TutorSerializer
+from rest_framework import generics
+
+class TutorList(generics.ListAPIView):
+    queryset = Tutor.objects.all()
+    serializer_class = TutorSerializer
+
+class TutorDetails(generics.RetrieveAPIView):
+    queryset = Tutor.objects.all()
+    serializer_class = TutorSerializer
