+++ backend/swpp/view/user.py
@@ -1,0 +1,12 @@
+from django.shortcuts import render
+from swpp.models import User
+from swpp.serializers import UserSerializer
+from rest_framework import generics
+
+class UserList(generics.ListAPIView):
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
+
+class UserDetails(generics.RetrieveAPIView):
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
